[elasticsearch] Fix and cleanup index creation
This commit fixes an issue with the index creation logic when the
elasticsearch.newdb flag is set to true. Namely, when running in local
mode the index would never exist (since the node always starts with a
clean temporary directory) but with elasticsearch.newdb an attempt would
be made to delete the index which would cause the workload to fail.

This commit also does a cleanup of the index creation logic using an
explicit create index request and passing the index settings along as
part of that index creation.