[elasticsearch] Return not found instead of error
This commit modifies the Elasticsearch client to return Status.NOT_FOUND
when an operation fails because the document requested does not
exist. This modifies the previous behavior which would return
Status.ERROR in this case.

This commit also moves the statements that would return Status.ERROR
into the corresponding catch blocks so that it is clear that
Status.ERROR is only being returned on error paths.