[couchbase2] Change query to use RAW and ORDER BY.
Using RAW queries reduces overhead at cbq-engine on the cluster and
reduces transport.  This in turn delivers lower latencies and thus
higher throughput with YCSB's tight loops of requests.

Adding an ORDER BY maintains strict correctness with the intended
implementation of Workload E.  In this particular case, the
ORDER BY is effectively a noop since the index in use is ordered
anyway.

The README for couchbase2 was also updated to reflect the simpler
index creation which is possible with Couchbase Server 4.5.  Server
4.5 automatically covers on the primary index so there is no need
for the more complicated syntax unless using an older version.

Note this change requires the 2.3.1 Couchbase Java SDK and its
underlying core-io 1.3.1.  The core-io 1.3.1 adds a streaming
parser for RAW results.